# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E_Messe API",
        "description": "API pour l'application E_Messe - Gestion des messes",
        "contact": {
            "email": "leprodev03@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Serveur Local"
        },
        {
            "url": "https://votre-domaine.com/api",
            "description": "Serveur de Production"
        }
    ],
    "paths": {
...
        "/messes": {
            "get": {
                "tags": [
                    "Messes"
                ],
                "summary": "Liste des messes actives de l'utilisateur connecté",
                "operationId": "dae7d67a90e96b3fc7af175acf4b5809",
                "responses": {
                    "200": {
                        "description": "Liste des messes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "messes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Messe"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messes"
                ],
                "summary": "Créer une nouvelle demande de messe",
                "operationId": "04d8e5af8aa792cc73f25397a7421f54",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motif_intention",
                                    "celebration_choisie",
                                    "montant_offrande",
                                    "date_souhaitee",
                                    "paroisse_id"
                                ],
                                "properties": {
                                    "motif_intention": {
                                        "type": "string"
                                    },
                                    "interception_par": {
                                        "type": "string"
                                    },
                                    "celebration_choisie": {
                                        "type": "string",
                                        "enum": [
                                            "Messe quotidienne",
                                            "Messe dominicale",
                                            "Messe solennelle"
                                        ]
                                    },
                                    "jours_quotidienne": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "jours_dominicale": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "montant_offrande": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "date_souhaitee": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "heure_souhaitee": {
                                        "type": "string",
                                        "format": "time"
                                    },
                                    "paroisse_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Messe créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "messe": {
                                            "$ref": "#/components/schemas/Messe"
                                        },
                                        "paiement": {
                                            "$ref": "#/components/schemas/Paiement"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messes/specifiques": {
            "get": {
                "tags": [
                    "Messes"
                ],
                "summary": "Liste des messes selon un type spécifique",
                "operationId": "104f45cf9787bb2daa18b61950ad3cbe",
                "parameters": [
                    {
                        "name": "celebration_choisie",
                        "in": "query",
                        "description": "Type de messe à filtrer",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Messe quotidienne",
                                "Messe dominicale",
                                "Messe solennelle"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste filtrée des messes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "celebration_choisie": {
                                            "type": "string"
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Messe"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Paramètre manquant"
                    },
                    "422": {
                        "description": "Type de célébration invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messes/history": {
            "get": {
                "tags": [
                    "Messes"
                ],
                "summary": "Historique des messes (annulées ou célébrées)",
                "operationId": "a3d3c6c6271302cd574a3f2c526cddb5",
                "responses": {
                    "200": {
                        "description": "Historique des messes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "history": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Messe"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messes/{messeId}": {
            "get": {
                "tags": [
                    "Messes"
                ],
                "summary": "Afficher les détails d’une messe spécifique",
                "operationId": "3461b99931889c28ef23587e42adfad3",
                "parameters": [
                    {
                        "name": "messeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la messe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Messe"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Messes"
                ],
                "summary": "Mettre à jour une demande de messe",
                "operationId": "34fbc5876e6496e11f8981d1fbe61ada",
                "parameters": [
                    {
                        "name": "messeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "motif_intention": {
                                        "type": "string"
                                    },
                                    "date_souhaitee": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour réussie"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "409": {
                        "description": "Cette demande ne peut plus être modifiée"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Messes"
                ],
                "summary": "Supprime une messe si non célébrée",
                "operationId": "e607900e03f0457363501d89754916eb",
                "parameters": [
                    {
                        "name": "messeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messe supprimée"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "409": {
                        "description": "Seules les messes en attente peuvent être supprimées"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Messe": {
                "title": "Messe",
                "description": "Représente une demande de messe",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "paroisse_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "motif_intention": {
                        "type": "string",
                        "example": "Pour la santé de ma famille"
                    },
                    "date_souhaitee": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-15"
                    },
                    "heure_souhaitee": {
                        "type": "string",
                        "format": "time",
                        "example": "09:30"
                    },
                    "celebration_choisie": {
                        "type": "string",
                        "example": "Messe dominicale"
                    },
                    "nom_demandeur": {
                        "type": "string",
                        "example": "Jean Dupont"
                    },
                    "email_demandeur": {
                        "type": "string",
                        "example": "jean@example.com"
                    },
                    "telephone_demandeur": {
                        "type": "string",
                        "example": "+2250700000000"
                    },
                    "montant_offrande": {
                        "type": "number",
                        "format": "float",
                        "example": 5000
                    },
                    "statut": {
                        "type": "string",
                        "example": "en_attente_paiement"
                    },
                    "dates_selectionnees": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Lundi"
                        }
                    },
                    "download_count": {
                        "type": "integer",
                        "example": 0
                    },
                    "last_downloaded_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-01T08:00:00Z"
                    },
                    "interception_par": {
                        "type": "string",
                        "example": "Père Martin"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}